# openapi.json
{
    "openapi": "3.0.1",
    "info": {
        "title": "API MiniMarket Plus",
        "description": "Microservicio backend de MiniMarket Plus. Expone la gestion de\nproductos, categorias, inventario, carrito de compras, ventas,\ndetalle de ventas y usuarios del sistema.\n\nEsta documentacion se genera con OpenAPI Specification (OAS 3)\nmediante springdoc-openapi, asegurando la consistencia entre el\ncodigo y la documentacion de la API.",
        "contact": {
            "name": "Equipo Backend - MiniMarket Plus",
            "email": "soporte@minimarketplus.cl"
        },
        "license": {
            "name": "Uso academico Duoc UC",
            "url": "https://www.duoc.cl"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8080",
            "description": "Entorno de desarrollo local"
        }
    ],
    "security": [
        {
            "autenticacionBasica": []
        }
    ],
...
    "paths": {
...
        "/api/carrito/{id}": {
            "get": {
                "tags": [
                    "Carrito"
                ],
                "summary": "Obtener item del carrito por ID",
                "description": "Devuelve la informacion detallada de un item del carrito segun su identificador.",
                "operationId": "obtenerCarritoPorId",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Identificador del item del carrito",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        },
                        "example": 1
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Item del carrito encontrado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Carrito"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "No autenticado"
                    },
                    "404": {
                        "description": "Item del carrito no encontrado"
                    }
                }
            },
            "put": {
                "tags": [
                    "Carrito"
                ],
                "summary": "Actualizar item del carrito",
                "description": "Actualiza los datos de un item existente del carrito identificado por su ID.",
                "operationId": "actualizarCarrito",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Identificador del item del carrito a actualizar",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        },
                        "example": 1
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Carrito"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Item del carrito actualizado correctamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Carrito"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "No autenticado"
                    },
                    "404": {
                        "description": "Item del carrito no encontrado"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Carrito"
                ],
                "summary": "Eliminar producto del carrito",
                "description": "Elimina un item del carrito de compras por su identificador.",
                "operationId": "eliminarProductoDelCarrito",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Identificador del item del carrito a eliminar",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        },
                        "example": 1
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Item del carrito eliminado (sin contenido)"
                    },
                    "401": {
                        "description": "No autenticado"
                    },
                    "404": {
                        "description": "Item del carrito no encontrado"
                    }
                }
            }
        },
...
        "/api/carrito": {
            "get": {
                "tags": [
                    "Carrito"
                ],
                "summary": "Listar carrito",
                "description": "Obtiene todos los items del carrito de compras de MiniMarket Plus.",
                "operationId": "listarCarrito",
                "responses": {
                    "401": {
                        "description": "No autenticado"
                    },
                    "200": {
                        "description": "Lista de items del carrito obtenida correctamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Carrito"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Carrito"
                ],
                "summary": "Agregar producto al carrito",
                "description": "Agrega un producto con su cantidad al carrito de compras de un usuario en MiniMarket Plus.",
                "operationId": "agregarProductoAlCarrito",
                "requestBody": {
                    "description": "Datos del producto a agregar al carrito",
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Carrito"
                            },
                            "examples": {
                                "Item de carrito nuevo": {
                                    "description": "Item de carrito nuevo",
                                    "value": {
                                        "usuario": {
                                            "id": 1
                                        },
                                        "producto": {
                                            "id": 1
                                        },
                                        "cantidad": 3
                                    }
                                }
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Producto agregado al carrito correctamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Carrito"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "No autenticado"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "Rol": {
                "required": [
                    "nombre"
                ],
                "type": "object",
                "properties": {
                    "id": {
                        "type": "integer",
                        "description": "Identificador unico del rol.",
                        "format": "int64",
                        "readOnly": true,
                        "example": 1
                    },
                    "nombre": {
                        "type": "string",
                        "description": "Nombre del rol segun convencion de Spring Security.",
                        "example": "ROLE_ADMIN"
                    }
                },
                "description": "Rol de autorizacion que agrupa permisos de acceso (RBAC)."
            },
            "Usuario": {
                "required": [
                    "password",
                    "username"
                ],
                "type": "object",
                "properties": {
                    "id": {
                        "type": "integer",
                        "description": "Identificador unico del usuario.",
                        "format": "int64",
                        "readOnly": true,
                        "example": 1
                    },
                    "username": {
                        "type": "string",
                        "description": "Nombre de usuario unico para autenticarse.",
                        "example": "admin"
                    },
                    "password": {
                        "type": "string",
                        "description": "Contrasena del usuario (se almacena cifrada con BCrypt). Solo escritura.",
                        "writeOnly": true,
                        "example": "admin123"
                    },
                    "roles": {
                        "uniqueItems": true,
                        "type": "array",
                        "description": "Roles asignados al usuario (por ejemplo ROLE_ADMIN).",
                        "items": {
                            "$ref": "#/components/schemas/Rol"
                        }
                    }
                },
                "description": "Usuario del sistema con sus credenciales y roles de acceso."
            },
            "Categoria": {
                "required": [
                    "nombre"
                ],
                "type": "object",
                "properties": {
                    "id": {
                        "type": "integer",
                        "description": "Identificador unico de la categoria.",
                        "format": "int64",
                        "readOnly": true,
                        "example": 1
                    },
                    "nombre": {
                        "type": "string",
                        "description": "Nombre unico de la categoria.",
                        "example": "Bebidas"
                    },
                    "productos": {
                        "type": "array",
                        "description": "Productos asociados a esta categoria.",
                        "readOnly": true,
                        "items": {
                            "$ref": "#/components/schemas/Producto"
                        }
                    }
                },
                "description": "Categoria que agrupa productos del catalogo (por ejemplo Bebidas o Abarrotes)."
            },
            "Producto": {
                "required": [
                    "categoria",
                    "nombre",
                    "precio",
                    "stock"
                ],
                "type": "object",
                "properties": {
                    "id": {
                        "type": "integer",
                        "description": "Identificador unico del producto.",
                        "format": "int64",
                        "readOnly": true,
                        "example": 1
                    },
                    "nombre": {
                        "type": "string",
                        "description": "Nombre comercial del producto.",
                        "example": "Agua Mineral 1L"
                    },
                    "precio": {
                        "type": "number",
                        "description": "Precio de venta unitario en pesos chilenos.",
                        "format": "double",
                        "example": 890.0
                    },
                    "stock": {
                        "type": "integer",
                        "description": "Cantidad disponible en stock.",
                        "format": "int32",
                        "example": 50
                    },
                    "categoria": {
                        "$ref": "#/components/schemas/Categoria"
                    }
                },
                "description": "Producto disponible en el catalogo de MiniMarket Plus."
            },
...
            "Carrito": {
                "required": [
                    "cantidad",
                    "producto",
                    "usuario"
                ],
                "type": "object",
                "properties": {
                    "id": {
                        "type": "integer",
                        "description": "Identificador unico del item del carrito.",
                        "format": "int64",
                        "readOnly": true,
                        "example": 1
                    },
                    "usuario": {
                        "$ref": "#/components/schemas/Usuario"
                    },
                    "producto": {
                        "$ref": "#/components/schemas/Producto"
                    },
                    "cantidad": {
                        "type": "integer",
                        "description": "Cantidad del producto solicitada.",
                        "format": "int32",
                        "example": 3
                    }
                },
                "description": "Item del carrito de compras: asocia un producto y una cantidad a un usuario."
            }
        },
        "securitySchemes": {
            "autenticacionBasica": {
                "type": "http",
                "description": "Autenticacion HTTP Basic. Use el usuario y la contrasena semilla (admin / admin123) en el boton Authorize para consumir los endpoints protegidos.",
                "name": "autenticacionBasica",
                "scheme": "basic"
            }
        }
    }
}